requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.24
django-environ==0.12.0
djangorestframework==3.16.1
idna==3.10
mysql-connector-python==9.4.0
parameterized==0.9.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
